requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==1.11.27
django-rest-framework==0.1.0
djangorestframework==3.9.4
idna==2.8
PyPDF2==1.26.0
pytz==2019.3
requests==2.22.0
urllib3==1.25.7
